Move search logic & HTML into SearchContainer
Move all the search logic and HTML elements implemented in the component
`App` to the `SearchContainer` component, ensuring the
`separation of concerns` design pattern is followed, making the code
cleaner to read
